TWorld: fixes after first hardware test
- Fix wrong frame delay comparison causing low FPS.
- Fix missing cast in LZSS 2 bytes backref decoding.
- Fix level pack rendering showing one extra pack when less than 4 packs.

Core library:
- Fix annoying clang-tidy warning asking to use lround in millis_to_ticks.
- Add FPS monitor utility.

At this point a 1 actor grid results in max 16.2 FPS, there is still optimization to do!

+++ sw/app/tworld/src/lzss.c
@@ -70,7 +70,7 @@ void lzss_decode(flash_t src, uint16_t length, void* dst) {
             const uint8_t *ref = out;
             if (b & 0x1) {
                 // two bytes encoding
-                uint16_t backref = (b | buf[buf_pos++] << 8) >> 1;
+                uint16_t backref = (uint16_t) (b | buf[buf_pos++] << 8) >> 1;
                 reflen = (backref & LENGTH_MASK2) + BREAKEVEN2;
                 ref -= backref >> LENGTH_BITS2;
                 --length;
